Restructure code to be single runnable script
